refactor(broker): extract shared PostgresStreamBase for filter/paginate/subscribe
PostgresMessageStream and PostgresEventStream duplicated the same
filter/paginate/subscribe/save boilerplate verbatim, which tripped
SonarQube's new-code duplication gate. Only append/all/delete/
getCurrentSequence differ (table name and columns), so those stay on
the concrete classes.
