+++ services/ark-broker/ark-broker/src/brokers/stream/postgres-message-stream.ts
@@ -1,16 +1,11 @@
-import {EventEmitter} from 'events';
 import type postgres from 'postgres';
 import type {Logger} from '@ark-broker/logging/logger.js';
 import type {Db} from '@ark-broker/db/db.js';
 import type {MessageData} from '../memory-broker.js';
 import {BrokerItem} from './broker-item.js';
-import {
-  DEFAULT_LIMIT,
-  type PaginatedList,
-  type PaginationParams,
-} from '../pagination.js';
 import type {Predicate} from './stream.js';
 import type {MessageStream} from './message-stream.js';
+import {PostgresStreamBase} from './postgres-stream-base.js';
 
 type MessageRow = {
   sequence_number: string;
@@ -32,14 +27,17 @@ function rowToBrokerItem(row: MessageRow): BrokerItem<MessageData> {
   };
 }
 
-export class PostgresMessageStream implements MessageStream {
-  private readonly emitter = new EventEmitter();
-
+export class PostgresMessageStream
+  extends PostgresStreamBase<MessageData>
+  implements MessageStream
+{
   constructor(
     private readonly logger: Logger,
     private readonly db: Db,
     private readonly ttlSeconds: number
-  ) {}
+  ) {
+    super();
+  }
 
   async append(
     data: MessageData,
@@ -71,38 +69,6 @@ export class PostgresMessageStream implements MessageStream {
     return rows.map(rowToBrokerItem);
   }
 
-  async filter(
-    predicate: Predicate<MessageData>
-  ): Promise<BrokerItem<MessageData>[]> {
-    return (await this.all()).filter(predicate);
-  }
-
-  async paginate(
-    params: PaginationParams,
-    predicate?: Predicate<MessageData>
-  ): Promise<PaginatedList<BrokerItem<MessageData>>> {
-    const limit = params.limit ?? DEFAULT_LIMIT;
-    const cursor = params.cursor;
-
-    const all = await this.all();
-    let filtered = predicate ? all.filter(predicate) : all;
-    const total = filtered.length;
-
-    if (cursor !== undefined) {
-      filtered = filtered.filter((item) => item.sequenceNumber > cursor);
-    }
-
-    const items = filtered.slice(0, limit);
-    const hasMore = filtered.length > limit;
-
-    return {
-      items,
-      total,
-      hasMore,
-      nextCursor: hasMore ? items.at(-1)!.sequenceNumber : undefined,
-    };
-  }
-
   async delete(predicate?: Predicate<MessageData>): Promise<void> {
     if (!predicate) {
       this.logger.info('deleting all messages');
@@ -121,19 +87,10 @@ export class PostgresMessageStream implements MessageStream {
     await this.db`DELETE FROM messages WHERE query_id = ${queryId}`;
   }
 
-  async save(): Promise<void> {}
-
   async getCurrentSequence(): Promise<number> {
     const [{seq}] = await this.db<[{seq: string | null}]>`
       SELECT MAX(sequence_number) as seq FROM messages WHERE expires_at > now()
     `;
     return seq === null ? 0 : Number(seq);
-  }
-
-  subscribe(callback: (item: BrokerItem<MessageData>) => void): () => void {
-    this.emitter.on('item', callback);
-    return (): void => {
-      this.emitter.off('item', callback);
-    };
   }
 }
